+++ src/main/java/org/agty/sql/dialect/mysql/MySQL.java
@@ -615,7 +615,7 @@ public class MySQL implements Sql {
                             .setQuoteColumn(getQuoteColumn())
                             .setQuoteValue(getQuoteValue())
                             .setColumn(key)
-                            .setValue(arguments.getFromData(key))
+                            .setValue(arguments.getData(key))
                             .setNoStringEncode(arguments.noStringEncode())
                             .build()
             );